% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generate.abstracts}
\alias{generate.abstracts}
\title{Function that generates a separate TeX file for each of the abstracts}
\usage{
generate.abstracts(my.filename, dirAbstracts, author.lastname = seq(23, by =
  6, length.out = 7), author.firstname = seq(24, by = 6, length.out = 7),
  author.institution = seq(26, by = 6, length.out = 7),
  author.city = seq(27, by = 6, length.out = 7), author.country = seq(3, by
  = 1, length.out = 7) + 1, author.email = seq(25, by = 6, length.out = 7),
  author.presenting = seq(22, by = 6, length.out = 7), pres.title = 18,
  pres.abstract = 21, accept = 17, topic1 = 19, topic2 = 3, id = 2,
  notes = 73, ref2 = 77, ref3 = 78, ref4 = 79, duplicated = NULL,
  accept.all = F, noNotes = F, style = "AS2012", notesDay = 81,
  notesPayment = 82, verbose = FALSE)
}
\arguments{
\item{my.filename}{tab delimited database that contains the abstract submission (contribution data base), full path must be given or it must be stored in the working directory of R}

\item{dirAbstracts}{directory where the abstracts will be stored, it must exist}

\item{author.lastname}{columns that contain the last name of the authors in the contribution data base (numbers must be gives, i.e. if only two authors are allowed and the last names are stored in the 5th and 10th column of the contribution database, author.lastname=c(5,10))}

\item{author.firstname}{columns that contain thefirst name of the authors}

\item{author.institution}{columns that contain the institution of the authors}

\item{author.city}{columns that contain the city of the authors}

\item{author.country}{columns that contain the country of the authors}

\item{author.email}{columns that contain the e-mails of the authors}

\item{author.presenting}{columns that contain an indicator T/F indicating if the author will present the work}

\item{pres.title}{column containing the title of the presentation}

\item{pres.abstract}{column containing the abstract text}

\item{accept}{column that contains the acceptance decision, must be (exaclty) "Yes" for the accepted abstracts}

\item{topic1}{column that contains the first selected topic}

\item{topic2}{column that contains the second selected topic}

\item{id}{column that contains the abstract ID}

\item{notes}{column that contains the notes added by the reviewers}

\item{ref2}{opinion of reviewer2}

\item{ref3}{opinion of reviewer3}

\item{ref4}{opinion of reviewer4}

\item{accept.all}{if set to TRUE, all the abstracts included in the database are "accepted", useful to generate the first draft of the abstract book}

\item{noNotes}{don't write notes, set to TRUE if notes should be omitted, useful for the version published for authors}

\item{style}{"AS2012" (used for the book of abstracts of Applied Statistics 2012) or "AS2011" (used for the book of abstracts of Applied Statistics 2011), the styles differ in how the e-mails and affiliations of the authors are formatted}

\item{notesDay}{column that contains the notes of the organizers about the (possible) day of presentation,}

\item{notesPayment}{column that contains the notes of the organizers about the payment/registration information about the presentation}

\item{verbose}{FALSE: indicates if some output should be written on screen while executing the function}
}
\value{
TeX files are generated and stored dirAbstracts directory, My.data: data set with submitted data; PA: index indicating the presenting author for each submission.
}
\description{
Function that generates a separate TeX file for each of the abstracts
}
\examples{
set.seed(1)
}

