% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generate.duties}
\alias{generate.duties}
\title{Function that generates the duties of the participants}
\usage{
generate.duties(my.filename, my.program.file, dirAbstracts,
  author.lastname = seq(23, by = 6, length.out = 7),
  author.firstname = seq(24, by = 6, length.out = 7),
  author.presenting = seq(22, by = 6, length.out = 7), accept = 17,
  id = 2, abstract.id = 10:14)
}
\arguments{
\item{my.filename:}{tab delimited database that contains the abstract submission}

\item{my.filename.program:}{: tab delimited database that contains the program of the conference}

\item{dirAbstracts:}{directory where the abstracts will be stored, it must exist}

\item{author.lastname:}{columns that contain the last name of the authors}

\item{author.firstname:}{columns that contain thefirst name of the authors}

\item{author.presenting:columns}{that contain an indicator T/F indicating if the author will present the work}

\item{accept:}{column that contains the acceptance decision, must be " Yes" for the acceptance of the abstract}

\item{id:}{column that contains the abstract ID}

\item{abstract.id:}{columns in the program file that contain the abstract IDs to be included in each session}
}
\value{
empty return; TeX files are generated and stored dirAbstracts directory
}
\description{
Function that generates the duties of the participants
}
\examples{
set.seed(1)
}

