% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generate.files.byDecision}
\alias{generate.files.byDecision}
\title{Function that writes in a directory a separate file containing the submissions for each decision}
\usage{
generate.files.byDecision(my.data, dirData, accept = 17,
  author.lastname = seq(23, by = 6, length.out = 7),
  author.firstname = seq(24, by = 6, length.out = 7), author.email = seq(25,
  by = 6, length.out = 7), presenting.author, file.names = "Submissions")
}
\arguments{
\item{my.data}{name of the R data.frame object containing the submissions, can be retrieved from the output obtained from the generate.abstracts function (first output)}

\item{dirData}{directory where the files will be stored, it must exist}

\item{accept}{column in the data.frame that contains the decision}

\item{author.lastname}{columns that contain the last name of the authors in the contribution data base (numbers must be gives, i.e. if only two authors are allowed and the last names are stored in the 5th and 10th column of the contribution database, author.lastname=c(5,10))}

\item{author.firstname}{columns that contain thefirst name of the authors}

\item{author.email}{columns that contain the e-mails of the authors}

\item{presenting.author}{index that indicates the order of the presenting author for each submission, can be retrieved from the output obtained from the generate.abstracts function (second output)}

\item{file.names}{prefix of the files names that will be generated, default is Submissions.The suffix will be given by the values included in the accept column}
}
\value{
The names of the files being generated. The functions writes tab-delimted text files in the dirData directory. Adds additional columns indicating the name and last name of the presenting author, and his/her e-mail.
}
\description{
Function that writes in a directory a separate file containing the submissions for each decision
}
\examples{
set.seed(1)
}

