% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generate.program}
\alias{generate.program}
\title{Function that generates the program of the conference}
\usage{
generate.program(my.program.file, my.filename, dirAbstracts,
  abstract.id = c(10:14), author.lastname = seq(23, by = 6, length.out = 7),
  author.firstname = seq(24, by = 6, length.out = 7), accept = 17,
  pres.title = 18, id = 75, outfile.name = "program.tex")
}
\arguments{
\item{my.program.file:}{name of the tab-delimited file where the program is saved}

\item{my.filename:}{name of the file with the abstract database}

\item{dirAbstracts:}{directory where the abstracts will be stored, it must exist}

\item{abstract.id:}{colums where the abstracts to be included in each session are reported}

\item{author.lastname:}{columns that contain the last name of the authors}

\item{author.first:}{columns that contain the first name of the authors}

\item{accept:}{column that contains the acceptance decision, must be " Yes" for the acceptance of the abstract}

\item{pres.title:}{column that contains the title of the presentation}

\item{id:}{abstract id column in the database of abstracts}
}
\value{
empty return; TeX files are generated and stored dirAbstracts directory
}
\description{
Function that generates the program of the conference
}
\examples{
set.seed(1)
}

