% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generate.programOverview}
\alias{generate.programOverview}
\title{Function that generates the program overview file}
\usage{
generate.programOverview(my.program.file, dirAbstracts,
  abstract.id = c(10:14))
}
\arguments{
\item{my.program.file:}{name of the tab-delimited file where the program is saved}

\item{dirAbstracts:}{directory where the abstracts will be stored, it must exist}

\item{abstract.id:}{colums where the abstracts to be included in each session are reported}
}
\description{
Function that generates the program overview file
}
\examples{
set.seed(1)
}

